+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Changed some members to readonly to fix SonarCloud's code smells.
+
 ## [2.8.1] - 2025-01-23
 
 ### Added