+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add GraphQL API documentation
 
+## [2.9.0] - 2025-01-23
+
+### Fixed
+- Changed some members to readonly to fix SonarCloud's code smells.
+
 ## [2.8.1] - 2025-01-23
 
 ### Added